+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/FormatExtension.java
@@ -18,8 +18,10 @@ package com.diffplug.gradle.spotless;
 import static com.diffplug.gradle.spotless.PluginGradlePreconditions.requireElementsNonNull;
 
 import java.io.File;
+import java.io.IOException;
 import java.io.Serializable;
 import java.nio.charset.Charset;
+import java.nio.file.Files;
 import java.util.*;
 import java.util.stream.Stream;
 
@@ -406,9 +408,15 @@ public class FormatExtension {
 		addStep(IndentStep.Type.TAB.create());
 	}
 
+	/**
+	 * Created by {@link FormatExtension#licenseHeader(String, String)} or {@link FormatExtension#licenseHeaderFile(Object, String)}.
+	 * For most language-specific formats (e.g. java, scala, etc.) you can omit the second `delimiter` argument, because it is supplied
+	 * automatically ({@link HasBuiltinDelimiterForLicense}).
+	 */
 	public abstract class LicenseHeaderConfig {
 		String delimiter;
 		String yearSeparator = LicenseHeaderStep.defaultYearDelimiter();
+		Boolean updateYearWithLatest = null;
 
 		public LicenseHeaderConfig(String delimiter) {
 			this.delimiter = Objects.requireNonNull(delimiter, "delimiter");
@@ -434,11 +442,29 @@ public class FormatExtension {
 			return this;
 		}
 
-		abstract FormatterStep createStep();
+		/**
+		 * @param updateYearWithLatest
+		 *           Will turn `2004` into `2004-2020`, and `2004-2019` into `2004-2020`
+		 *           Default value is false, unless {@link SpotlessExtension#ratchetFrom(String)} is used, in which case default value is true.
+		 */
+		public LicenseHeaderConfig updateYearWithLatest(boolean overwriteYearLatest) {
+			this.updateYearWithLatest = overwriteYearLatest;
+			replaceStep(createStep());
+			return this;
+		}
+
+		protected abstract String licenseHeader() throws IOException;
+
+		FormatterStep createStep() {
+			return FormatterStep.createLazy(LicenseHeaderStep.name(), () -> {
+				// by default, we should update the year if the user is using ratchetFrom
+				boolean updateYear = updateYearWithLatest == null ? FormatExtension.this.root.getRatchetFrom() != null : updateYearWithLatest;
+				return new LicenseHeaderStep(licenseHeader(), delimiter, yearSeparator, updateYear);
+			}, step -> step::format);
+		}
 	}
 
-	public class LicenseStringHeaderConfig extends LicenseHeaderConfig {
-
+	private class LicenseStringHeaderConfig extends LicenseHeaderConfig {
 		private String header;
 
 		LicenseStringHeaderConfig(String delimiter, String header) {
@@ -446,13 +472,13 @@ public class FormatExtension {
 			this.header = Objects.requireNonNull(header, "header");
 		}
 
-		FormatterStep createStep() {
-			return LicenseHeaderStep.createFromHeader(header, delimiter, yearSeparator);
+		@Override
+		protected String licenseHeader() {
+			return header;
 		}
 	}
 
-	public class LicenseFileHeaderConfig extends LicenseHeaderConfig {
-
+	private class LicenseFileHeaderConfig extends LicenseHeaderConfig {
 		private Object headerFile;
 
 		LicenseFileHeaderConfig(String delimiter, Object headerFile) {
@@ -460,10 +486,10 @@ public class FormatExtension {
 			this.headerFile = Objects.requireNonNull(headerFile, "headerFile");
 		}
 
-		FormatterStep createStep() {
-			return LicenseHeaderStep
-					.createFromFile(getProject().file(headerFile), getEncoding(), delimiter,
-							yearSeparator);
+		@Override
+		protected String licenseHeader() throws IOException {
+			byte[] content = Files.readAllBytes(getProject().file(headerFile).toPath());
+			return new String(content, getEncoding());
 		}
 	}